+++ code/spikes/publish.001/package.json
@@ -28,8 +28,8 @@
   "private": true,
   "types": "./dist/types.d/index.html.d.html",
   "exports": {
-    ".": "./dist/web/assets/index.49869af2.js",
-    "./web": "./dist/web/assets/index.49869af2.js"
+    ".": "./dist/web/assets/index.469b59ce.js",
+    "./web": "./dist/web/assets/index.469b59ce.js"
   },
   "typesVersions": {
     "*": {
