+++ code/spikes/publish.001/src/ui.react/History/HistoryUtil.mts
