+++ code/spikes/publish.001/src/ui.react/common/types.mts
